refactor(decisions): consolidate proposal editor route files
Move logic from (editor)/edit/layout.tsx into the page itself, since
the editor route has no children that would benefit from a persistent
layout. Restore loading.tsx to (editor)/ so it actually wraps the page
(the prior location at [profileId]/edit/loading.tsx sat in a parallel
folder branch and never triggered).
