--- apps/app/src/app/[locale]/(no-header)/decisions/[slug]/proposal/[profileId]/edit/loading.tsx
@@ -1,5 +1,0 @@
-import { ProposalEditorSkeleton } from '@/components/decisions/ProposalEditorSkeleton';
-
-export default function Loading() {
-  return <ProposalEditorSkeleton />;
-}
